+++ CHANGELOG.md
@@ -19,7 +19,7 @@ If you like this project and find it useful, please consider giving it a star on
 - [package]: Updated package.
 - [package]: Updated dependencies.
 - [plugin]: Requires Matterbridge 3.0.3.
-- [config]: Removed parameter postfixHostname. Use postfix if needed.
+- [config]: As anticipated in the previous release, removed the parameter postfixHostname. Use postfix if needed.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="bmc-button.svg" alt="Buy me a coffee" width="80">